[TSENS] Regenerated demo for tsens__tsens_temp_measurement__firmware__sam_c21n_xpro.X
MH3-101478

+++ apps/tsens/tsens_temp_measurement/firmware/src/config/sam_c21n_xpro/peripheral/tsens/plib_tsens.c
@@ -136,7 +136,7 @@ void TSENS_ConversionStart( void )
 
 
 /* Configure window comparison threshold values */
-void TSENS_ComparisonWindowSet(uint16_t low_threshold, uint16_t high_threshold)
+void TSENS_ComparisonWindowSet(uint32_t low_threshold, uint32_t high_threshold)
 {
     TSENS_REGS->TSENS_WINLT = low_threshold;
     TSENS_REGS->TSENS_WINUT = high_threshold;
